Add portal command handlers for starting, stopping, and listing portals
