+++ include/core/cli-handlers/portal-cli-handler.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void handle_start_portal(int argc, char **argv);
+void handle_stop_portal(int argc, char **argv);
+void handle_listportals(int argc, char **argv);
